[FP8][Config] Use uint8 packing for fp8_e4m3_e4m3_max (#248)

+++ python/mlc_chat/quantization/quantization.py
@@ -197,7 +197,7 @@ QUANTIZATION: Dict[str, Quantization] = {
         kind="per-tensor-quant",
         activation_dtype="e4m3_float8",
         weight_dtype="e4m3_float8",
-        storage_dtype="uint32",
+        storage_dtype="uint8",
         model_dtype="float16",
         quantize_embedding=False,
     ),
